refactor(hickey): rename deploy freshness boundary
Rename the stale pwa.ts module so deploy freshness and legacy cleanup no longer imply active PWA service-worker ownership.
